NFC: Add support for dummy FW recovery
eSE needs to be powered up to establish RoT
and HMAC on device boot-up which happens in early
boot-up stage, while the normal FW download/recovery
procedure is triggered later when the HLOS has booted
up and NFC java service is up and running.
HMAC establishment would result into failure
and the device boot-up would fail as the FW
would never be recovered.
Dummy FW recovery would be done in order to
power up the NFC/eSE if FW was found in
tear-down mode in order to power-up the eSE
during early Android boot-up.
Normal FW recovery would still be triggered later
on when NFC java service would be up and running.

Change-Id: I90aeb4dd62abfbdb128918a476b8da575566e1b8